consumers/requirements.txt
influxdb-client==1.39.0
findspark==2.0.1 
pyspark==3.5.0
python-dotenv==1.0.0